ci: Yolorework i18n tests, add a maintenance/ folder
