+++ .gitlab/ci/test.gitlab-ci.yml
@@ -23,7 +23,6 @@
         artifacts: true
       - job: upgrade
 
-
 ########################################
 # TESTS
 ########################################
@@ -50,24 +49,6 @@ full-tests:
   artifacts:
     reports:
       junit: report.xml
-
-test-i18n-keys:
-  extends: .test-stage
-  script:
-    - python3 -m pytest tests/test_i18n_keys.py
-  only:
-    changes:
-      - locales/en.json
-      - src/*.py
-      - src/diagnosers/*.py
-
-test-translation-format-consistency:
-  extends: .test-stage
-  script:
-    - python3 -m pytest tests/test_translation_format_consistency.py
-  only:
-    changes:
-      - locales/*
 
 test-actionmap:
   extends: .test-stage
