+++ .gitlab/ci/translation.gitlab-ci.yml
@@ -1,6 +1,14 @@
 ########################################
 # TRANSLATION
 ########################################
+test-i18n-keys:
+  script:
+    - python3 maintenance/missing_i18n_keys --check
+  only:
+    changes:
+      - locales/en.json
+      - src/*.py
+      - src/diagnosers/*.py
 
 autofix-translated-strings:
   stage: translation
@@ -12,12 +20,10 @@ autofix-translated-strings:
     - git config --global user.name "$GITHUB_USER"
     - git remote set-url origin https://$GITHUB_TOKEN:x-oauth-basic@github.com/YunoHost/yunohost.git
   script:
-    - cd tests # Maybe move this script location to another folder?
     # create a local branch that will overwrite distant one
     - git checkout -b "ci-autofix-translated-strings-${CI_COMMIT_REF_NAME}" --no-track
-    - python3 remove_stale_translated_strings.py
-    - python3 autofix_locale_format.py
-    - python3 reformat_locales.py
+    - python3 maintenance/missing_i18n_keys --fix
+    - python3 maintenanceautofix_locale_format.py
     - '[ $(git diff -w | wc -l) != 0 ] || exit 0'  # stop if there is nothing to commit
     - git commit -am "[CI] Reformat / remove stale translated strings" || true
     - git push -f origin "HEAD":"ci-remove-stale-translated-strings-${CI_COMMIT_REF_NAME}"
