+++ locales/eo.json
@@ -51,7 +51,6 @@
     "ask_new_path": "Nova vojo",
     "backup_custom_mount_error": "Propra rezerva metodo ne povis preterpasi la paŝon 'monto'",
     "app_upgrade_app_name": "Nun ĝisdatigu {app}...",
-    "app_manifest_invalid": "Io misas pri la aplika manifesto: {error}",
     "backup_cleaning_failed": "Ne povis purigi la provizoran rezervan dosierujon",
     "backup_creation_failed": "Ne povis krei la rezervan ar archiveivon",
     "backup_hook_unknown": "La rezerva hoko '{hook}' estas nekonata",
@@ -219,7 +218,6 @@
     "pattern_mailbox_quota": "Devas esti grandeco kun la sufikso b/k/M/G/T aŭ 0 por ne havi kvoton",
     "user_deletion_failed": "Ne povis forigi uzanton {user}: {error}",
     "backup_with_no_backup_script_for_app": "La app '{app}' ne havas sekretan skripton. Ignorante.",
-    "service_regen_conf_is_deprecated": "'yunohost service regen-conf' malakceptas! Bonvolu uzi anstataŭe 'yunohost tools regen-conf'.",
     "global_settings_key_doesnt_exists": "La ŝlosilo '{settings_key}' ne ekzistas en la tutmondaj agordoj, vi povas vidi ĉiujn disponeblajn klavojn per uzado de 'yunohost settings list'",
     "dyndns_no_domain_registered": "Neniu domajno registrita ĉe DynDNS",
     "dyndns_could_not_check_available": "Ne povis kontroli ĉu {domain} haveblas sur {provider}.",
@@ -235,16 +233,13 @@
     "service_stop_failed": "Ne povis maldaŭrigi la servon '{service}'\n\nLastatempaj servaj protokoloj: {logs}",
     "unbackup_app": "App '{app}' ne konserviĝos",
     "updating_apt_cache": "Akirante haveblajn ĝisdatigojn por sistemaj pakoj…",
-    "tools_upgrade_at_least_one": "Bonvolu specifi '--apps' aŭ '--system'",
     "service_already_stopped": "La servo '{service}' jam ĉesis",
-    "tools_upgrade_cant_both": "Ne eblas ĝisdatigi ambaŭ sistemon kaj programojn samtempe",
     "restore_extracting": "Eltirante bezonatajn dosierojn el la ar theivo…",
     "upnp_port_open_failed": "Ne povis malfermi havenon per UPnP",
     "log_app_upgrade": "Ĝisdatigu la aplikon '{}'",
     "log_help_to_get_failed_log": "La operacio '{desc}' ne povis finiĝi. Bonvolu dividi la plenan ŝtipon de ĉi tiu operacio per la komando 'yunohost log share {name}' por akiri helpon",
     "port_already_closed": "Haveno {port} estas jam fermita por {ip_version} rilatoj",
     "hook_name_unknown": "Nekonata hoko-nomo '{name}'",
-    "dyndns_could_not_check_provide": "Ne povis kontroli ĉu {provider} povas provizi {domain}.",
     "restore_nothings_done": "Nenio estis restarigita",
     "log_tools_postinstall": "Afiŝu vian servilon YunoHost",
     "dyndns_unavailable": "La domajno '{domain}' ne haveblas.",
