+++ locales/es.json
@@ -10,7 +10,6 @@
     "app_extraction_failed": "No se pudieron extraer los archivos de instalación",
     "app_id_invalid": "ID de la aplicación no válida",
     "app_install_files_invalid": "Estos archivos no se pueden instalar",
-    "app_manifest_invalid": "Algo va mal con el manifiesto de la aplicación: {error}",
     "app_not_correctly_installed": "La aplicación {app} 8 parece estar incorrectamente instalada",
     "app_not_installed": "No se pudo encontrar «{app}» en la lista de aplicaciones instaladas: {all_apps}",
     "app_not_properly_removed": "La {app} 0 no ha sido desinstalada correctamente",
@@ -194,7 +193,6 @@
     "backup_system_part_failed": "No se pudo respaldar la parte del sistema «{part}»",
     "backup_with_no_backup_script_for_app": "La aplicación «{app}» no tiene un guión de respaldo. Omitiendo.",
     "backup_with_no_restore_script_for_app": "«{app}» no tiene un script de restauración, no podá restaurar automáticamente la copia de seguridad de esta aplicación.",
-    "dyndns_could_not_check_provide": "No se pudo verificar si {provider} puede ofrecer {domain}.",
     "dyndns_domain_not_provided": "El proveedor de DynDNS {provider} no puede proporcionar el dominio {domain}.",
     "experimental_feature": "Aviso : esta funcionalidad es experimental y no se considera estable, no debería usarla a menos que sepa lo que está haciendo.",
     "good_practices_about_user_password": "Ahora está a punto de definir una nueva contraseña de usuario. La contraseña debe tener al menos 8 caracteres, aunque es una buena práctica usar una contraseña más larga (es decir, una frase de contraseña) y / o una variación de caracteres (mayúsculas, minúsculas, dígitos y caracteres especiales).",
@@ -212,8 +210,6 @@
     "tools_upgrade_regular_packages": "Actualizando ahora paquetes «normales» (no relacionados con YunoHost)…",
     "tools_upgrade_cant_unhold_critical_packages": "No se pudo liberar los paquetes críticos…",
     "tools_upgrade_cant_hold_critical_packages": "No se pudieron retener los paquetes  críticos…",
-    "tools_upgrade_cant_both": "No se puede actualizar el sistema y las aplicaciones al mismo tiempo",
-    "tools_upgrade_at_least_one": "Especifique «--apps», o «--system»",
     "this_action_broke_dpkg": "Esta acción rompió dpkg/APT(los gestores de paquetes del sistema)…  Puede tratar de solucionar este problema conectando mediante SSH y ejecutando `sudo dpkg --configure -a`.",
     "service_reloaded_or_restarted": "El servicio '{service}' fue recargado o reiniciado",
     "service_reload_or_restart_failed": "No se pudo recargar o reiniciar el servicio «{service}»\n\nRegistro de servicios recientes:{logs}",
@@ -221,7 +217,6 @@
     "service_restart_failed": "No se pudo reiniciar el servicio «{service}»\n\nRegistro de servicios recientes:{logs}",
     "service_reloaded": "Servicio '{service}' recargado",
     "service_reload_failed": "No se pudo recargar el servicio «{service}»\n\nRegistro de servicios recientes:{logs}",
-    "service_regen_conf_is_deprecated": "¡«yunohost service regen-conf» está obsoleto! Use «yunohost tools regen-conf» en su lugar.",
     "service_description_yunohost-firewall": "Gestiona los puertos de conexiones abiertos y cerrados a los servicios",
     "service_description_yunohost-api": "Gestiona las interacciones entre la interfaz web de YunoHost y el sistema",
     "service_description_ssh": "Permite conectar a su servidor remotamente mediante un terminal (protocolo SSH)",
@@ -524,19 +519,6 @@
     "app_manifest_install_ask_domain": "Seleccione el dominio donde esta app debería ser instalada",
     "app_label_deprecated": "Este comando está depreciado! Favor usar el nuevo comando 'yunohost user permission update' para administrar la etiqueta de app.",
     "app_argument_password_no_default": "Error al interpretar argumento de contraseña'{name}': El argumento de contraseña no puede tener un valor por defecto por razón de seguridad",
-    "migration_0015_not_enough_free_space": "¡El espacio es muy bajo en `/var/`! Deberías tener almenos 1Gb de espacio libre para ejecutar la migración.",
-    "migration_0015_not_stretch": "¡La distribución actual de Debian no es Stretch!",
-    "migration_0015_yunohost_upgrade": "Iniciando la actualización del núcleo de YunoHost...",
-    "migration_0015_still_on_stretch_after_main_upgrade": "Algo fue mal durante la actualización principal, el sistema parece que está todavía en Debian Stretch",
-    "migration_0015_main_upgrade": "Comenzando la actualización principal...",
-    "migration_0015_patching_sources_list": "Adaptando las sources.lists...",
-    "migration_0015_start": "Comenzando la migración a Buster",
-    "migration_description_0019_extend_permissions_features": "Extiende/rehaz el sistema de gestión de permisos de la aplicación",
-    "migration_description_0018_xtable_to_nftable": "Migra las viejas reglas de tráfico de red al nuevo sistema nftable",
-    "migration_description_0017_postgresql_9p6_to_11": "Migra las bases de datos de PostgreSQL 9.6 a 11",
-    "migration_description_0016_php70_to_php73_pools": "Migra el «pool» de ficheros php7.0-fpm a php7.3",
-    "migration_description_0015_migrate_to_buster": "Actualiza el sistema a Debian Buster y YunoHost 4.x",
-    "migrating_legacy_permission_settings": "Migrando los antiguos parámetros de permisos...",
     "invalid_regex": "Regex no valido: «{regex}»",
     "global_settings_setting_backup_compress_tar_archives": "Cuando se creen nuevas copias de respaldo, comprimir los archivos (.tar.gz) en lugar de descomprimir los archivos (.tar). N.B.: activar esta opción quiere decir que los archivos serán más pequeños pero que el proceso tardará más y utilizará más CPU.",
     "global_settings_setting_smtp_relay_password": "Clave de uso del SMTP",
@@ -544,7 +526,6 @@
     "global_settings_setting_smtp_relay_port": "Puerto de envio / relay SMTP",
     "global_settings_setting_smtp_relay_host": "El servidor relay de SMTP para enviar correo en lugar de esta instalación YunoHost. Útil si estás en una de estas situaciones: tu puerto 25 esta bloqueado por tu ISP o VPS, si estás en usado una IP marcada como residencial o DUHL, si no puedes configurar un DNS inverso o si el servidor no está directamente expuesto a internet y quieres utilizar otro servidor para enviar correos.",
     "global_settings_setting_smtp_allow_ipv6": "Permitir el uso de IPv6 para enviar y recibir correo",
-    "domain_name_unknown": "Dominio «{domain}» desconocido",
     "diagnosis_processes_killed_by_oom_reaper": "Algunos procesos fueron terminados por el sistema recientemente porque se quedó sin memoria. Típicamente es sintoma de falta de memoria o de un proceso que se adjudicó demasiada memoria.<br>Resumen de los procesos terminados:<br>\n{kills_summary}",
     "diagnosis_http_nginx_conf_not_up_to_date_details": "Para arreglar este asunto, estudia las diferencias mediante el comando <cmd>yunohost tools regen-conf nginx --dry-run --with-diff</cmd> y si te parecen bien aplica los cambios mediante <cmd>yunohost tools regen-conf nginx --force</cmd>.",
     "diagnosis_http_nginx_conf_not_up_to_date": "Parece que la configuración nginx de este dominio haya sido modificada manualmente, esto no deja que YunoHost pueda diagnosticar si es accesible mediante HTTP.",
