+++ locales/oc.json
@@ -33,7 +33,6 @@
     "app_change_url_identical_domains": "L’ancian e lo novèl coble domeni/camin son identics per {domain}{path}, pas res a far.",
     "app_change_url_success": "L’URL de l’aplicacion {app} es ara {domain}{path}",
     "app_extraction_failed": "Extraccion dels fichièrs d’installacion impossibla",
-    "app_manifest_invalid": "I a quicòm que truca amb lo manifest de l’aplicacion : {error}",
     "app_requirements_checking": "Verificacion dels paquets requesits per {app}...",
     "app_sources_fetch_failed": "Recuperacion dels fichièrs fonts impossibla, l’URL es corrècta ?",
     "app_unsupported_remote_type": "Lo tipe alonhat utilizat per l’aplicacion es pas suportat",
@@ -110,7 +109,6 @@
     "domains_available": "Domenis disponibles :",
     "done": "Acabat",
     "downloading": "Telecargament…",
-    "dyndns_could_not_check_provide": "Impossible de verificar se {provider} pòt provesir {domain}.",
     "dyndns_ip_update_failed": "Impossible d’actualizar l’adreça IP sul domeni DynDNS",
     "dyndns_ip_updated": "Vòstra adreça IP actualizada pel domeni DynDNS",
     "dyndns_key_generating": "La clau DNS es a se generar… pòt trigar una estona.",
@@ -332,18 +330,15 @@
     "regenconf_dry_pending_applying": "Verificacion de la configuracion que seriá estada aplicada a la categoria « {category} »…",
     "regenconf_failed": "Regeneracion impossibla de la configuracion per la(s) categoria(s) : {categories}",
     "regenconf_pending_applying": "Aplicacion de la configuracion en espèra per la categoria « {category} »…",
-    "tools_upgrade_cant_both": "Actualizacion impossibla del sistèma e de las aplicacions a l’encòp",
     "tools_upgrade_cant_hold_critical_packages": "Manteniment impossible dels paquets critiques…",
     "global_settings_setting_security_nginx_compatibility": "Solucion de compromés entre compatibilitat e seguretat pel servidor web NGINX Afècta los criptografs (e d’autres aspèctes ligats amb la seguretat)",
     "global_settings_setting_security_ssh_compatibility": "Solucion de compromés entre compatibilitat e seguretat pel servidor SSH. Afècta los criptografs (e d’autres aspèctes ligats amb la seguretat)",
     "global_settings_setting_security_postfix_compatibility": "Solucion de compromés entre compatibilitat e seguretat pel servidor Postfix. Afècta los criptografs (e d’autres aspèctes ligats amb la seguretat)",
-    "service_regen_conf_is_deprecated": "« yunohost service regen-conf » es despreciat ! Utilizatz « yunohost tools regen-conf » allòc.",
     "service_reload_failed": "Impossible de recargar lo servici « {service} »\n\nJornal d’audit recent : {logs}",
     "service_restart_failed": "Impossible de reaviar lo servici « {service} »\n\nJornal d’audit recent : {logs}",
     "service_reload_or_restart_failed": "Impossible de recargar o reaviar lo servici « {service} »\n\nJornal d’audit recent : {logs}",
     "regenconf_file_kept_back": "S’espèra que lo fichièr de configuracion « {conf} » siá suprimit per regen-conf (categoria {category} mas es estat mantengut.",
     "this_action_broke_dpkg": "Aquesta accion a copat dpkg/apt (los gestionaris de paquets del sistèma)… Podètz ensajar de resòlver aqueste problèma en vos connectant amb SSH e executant « sudo dpkg --configure -a ».",
-    "tools_upgrade_at_least_one": "Especificatz --apps O --system",
     "tools_upgrade_cant_unhold_critical_packages": "Se pòt pas quitar de manténer los paquets critics…",
     "tools_upgrade_regular_packages": "Actualizacion dels paquets « normals » (pas ligats a YunoHost)…",
     "tools_upgrade_special_packages": "Actualizacion dels paquets « especials » (ligats a YunoHost)…",
@@ -480,7 +475,6 @@
     "diagnosis_basesystem_hardware": "L’arquitectura del servidor es {virt} {arch}",
     "backup_archive_corrupted": "Sembla que l’archiu de la salvagarda « {archive} » es corromput : {error}",
     "diagnosis_domain_expires_in": "{domain} expiraà d’aquí {days} jorns.",
-    "migration_0015_cleaning_up": "Netejatge de la memòria cache e dels paquets pas mai necessaris…",
     "restore_already_installed_apps": "Restauracion impossibla de las aplicacions seguentas que son ja installadas : {apps}",
     "diagnosis_package_installed_from_sury": "D’unes paquets sistèma devon èsser meses a nivèl",
     "ask_user_domain": "Domeni d’utilizar per l’adreça de corrièl de l’utilizaire e lo compte XMPP",
@@ -493,14 +487,6 @@
     "app_label_deprecated": "Aquesta comanda es estada renduda obsolèta. Mercés d'utilizar lo nòva \"yunohost user permission update\" per gerir letiquetada de l'aplication",
     "additional_urls_already_removed": "URL addicionala {url} es ja estada elimida per la permission «#permission:s»",
     "additional_urls_already_added": "URL addicionadal «{url}'» es ja estada aponduda per la permission «{permission}»",
-    "migration_0015_yunohost_upgrade": "Aviada de la mesa a jorn de YunoHost...",
-    "migration_0015_main_upgrade": "Aviada de la mesa a nivèl generala...",
-    "migration_0015_patching_sources_list": "Mesa a jorn del fichièr sources.lists...",
-    "migration_0015_start": "Aviar la migracion cap a Buster",
-    "migration_description_0017_postgresql_9p6_to_11": "Migrar las basas de donadas de PostgreSQL 9.6 cap a 11",
-    "migration_description_0016_php70_to_php73_pools": "Migrar los fichièrs de configuracion php7.0 cap a php7.3",
-    "migration_description_0015_migrate_to_buster": "Mesa a nivèl dels sistèmas Debian Buster e YunoHost 4.x",
-    "migrating_legacy_permission_settings": "Migracion dels paramètres de permission ancians...",
     "log_app_action_run": "Executar l’accion de l’aplicacion « {} »",
     "diagnosis_basesystem_hardware_model": "Lo modèl del servidor es {model}",
     "backup_archive_cant_retrieve_info_json": "Obtencion impossibla de las informacions de l’archiu « {archive} »... Se pòt pas recuperar lo fichièr info.json (o es pas un fichièr json valid).",
