+++ locales/zh_Hans.json
@@ -137,7 +137,6 @@
     "additional_urls_already_removed": "权限'{permission}'的其他URL中已经删除了附加URL'{url}'",
     "app_manifest_install_ask_path": "选择安装此应用的路径",
     "app_manifest_install_ask_domain": "选择应安装此应用程序的域",
-    "app_manifest_invalid": "应用清单错误: {error}",
     "app_location_unavailable": "该URL不可用，或与已安装的应用冲突：\n{apps}",
     "app_label_deprecated": "不推荐使用此命令！请使用新命令 'yunohost user permission update'来管理应用标签。",
     "app_make_default_location_already_used": "无法将'{app}' 设置为域上的默认应用，'{other_app}'已在使用'{domain}'",
@@ -218,7 +217,6 @@
     "service_description_rspamd": "过滤垃圾邮件和其他与电子邮件相关的功能",
     "service_description_redis-server": "用于快速数据访问，任务队列和程序之间通信的专用数据库",
     "service_description_postfix": "用于发送和接收电子邮件",
-    "service_description_php7.3-fpm": "使用NGINX运行用PHP编写的应用程序",
     "service_description_nginx": "为你的服务器上托管的所有网站提供服务或访问",
     "service_description_mysql": "存储应用程序数据（SQL数据库）",
     "service_description_metronome": "管理XMPP即时消息传递帐户",
@@ -235,7 +233,6 @@
     "service_reload_failed": "无法重新加载服务'{service}'\n\n最近的服务日志:{logs}",
     "service_removed": "服务 '{service}' 已删除",
     "service_remove_failed": "无法删除服务'{service}'",
-    "service_regen_conf_is_deprecated": "不建议使用'yunohost service regen-conf' ! 请改用'yunohost tools regen-conf'。",
     "service_enabled": "现在，服务'{service}' 将在系统引导过程中自动启动。",
     "service_enable_failed": "无法使服务 '{service}'在启动时自动启动。\n\n最近的服务日志:{logs}",
     "service_disabled": "系统启动时，服务 '{service}' 将不再启动。",
@@ -243,8 +240,6 @@
     "tools_upgrade_regular_packages": "现在正在升级 'regular' （与yunohost无关）的软件包…",
     "tools_upgrade_cant_unhold_critical_packages": "无法解压关键软件包…",
     "tools_upgrade_cant_hold_critical_packages": "无法保存重要软件包…",
-    "tools_upgrade_cant_both": "无法同时升级系统和应用程序",
-    "tools_upgrade_at_least_one": "请指定'apps', 或 'system'",
     "this_action_broke_dpkg": "此操作破坏了dpkg / APT（系统软件包管理器）...您可以尝试通过SSH连接并运行`sudo apt install --fix-broken`和/或`sudo dpkg --configure -a`来解决此问题。",
     "system_username_exists": "用户名已存在于系统用户列表中",
     "system_upgraded": "系统升级",
@@ -353,13 +348,11 @@
     "dyndns_ip_updated": "在DynDNS上更新了您的IP",
     "dyndns_ip_update_failed": "无法将IP地址更新到DynDNS",
     "dyndns_could_not_check_available": "无法检查{provider}上是否可用 {domain}。",
-    "dyndns_could_not_check_provide": "无法检查{provider}是否可以提供 {domain}.",
     "dpkg_lock_not_available": "该命令现在无法运行，因为另一个程序似乎正在使用dpkg锁（系统软件包管理器）",
     "dpkg_is_broken": "您现在不能执行此操作，因为dpkg / APT（系统软件包管理器）似乎处于损坏状态……您可以尝试通过SSH连接并运行sudo apt install --fix-broken和/或 sudo dpkg --configure-a 来解决此问题.",
     "downloading": "下载中…",
     "done": "完成",
     "domains_available": "可用域：",
-    "domain_name_unknown": "域'{domain}'未知",
     "domain_uninstall_app_first": "这些应用程序仍安装在您的域中：\n{apps}\n\n请先使用 'yunohost app remove the_app_id' 将其卸载，或使用 'yunohost app change-url the_app_id'将其移至另一个域，然后再继续删除域",
     "domain_remove_confirm_apps_removal": "删除该域将删除这些应用程序：\n{apps}\n\n您确定要这样做吗? [{answers}]",
     "domain_hostname_failed": "无法设置新的主机名。稍后可能会引起问题（可能没问题）。",
@@ -571,39 +564,10 @@
     "migrations_dependencies_not_satisfied": "在迁移{id}之前运行以下迁移: '{dependencies_id}'。",
     "migrations_cant_reach_migration_file": "无法访问路径'%s'处的迁移文件",
     "migrations_already_ran": "这些迁移已经完成: {ids}",
-    "migration_0019_slapd_config_will_be_overwritten": "好像您手动编辑了slapd配置。对于此关键迁移，YunoHost需要强制更新slapd配置。原始文件将备份在{conf_backup_folder}中。",
-    "migration_0019_add_new_attributes_in_ldap": "在LDAP数据库中添加权限的新属性",
-    "migration_0018_failed_to_reset_legacy_rules": "无法重置旧版iptables规则: {error}",
-    "migration_0018_failed_to_migrate_iptables_rules": "无法将旧的iptables规则迁移到nftables: {error}",
-    "migration_0017_not_enough_space": "在{path}中提供足够的空间来运行迁移。",
-    "migration_0017_postgresql_11_not_installed": "已安装PostgreSQL 9.6，但未安装PostgreSQL11？您的系统上可能发生了一些奇怪的事情:(...",
-    "migration_0017_postgresql_96_not_installed": "PostgreSQL未安装在您的系统上。无事可做。",
-    "migration_0015_weak_certs": "发现以下证书仍然使用弱签名算法，并且必须升级以与下一版本的nginx兼容: {certs}",
-    "migration_0015_cleaning_up": "清理不再有用的缓存和软件包...",
-    "migration_0015_specific_upgrade": "开始升级需要独立升级的系统软件包...",
-    "migration_0015_modified_files": "请注意，发现以下文件是手动修改的，并且在升级后可能会被覆盖: {manually_modified_files}",
-    "migration_0015_problematic_apps_warning": "请注意，已检测到以下可能有问题的已安装应用程序。看起来好像那些不是从YunoHost应用程序目录中安装的，或者没有标记为“正在运行”。因此，不能保证它们在升级后仍然可以使用: {problematic_apps}",
-    "migration_0015_general_warning": "请注意，此迁移是一项微妙的操作。YunoHost团队竭尽全力对其进行检查和测试，但迁移仍可能会破坏系统或其应用程序的某些部分。\n\n因此，建议：\n -对任何关键数据或应用程序执行备份。有关更多信息，请访问https://yunohost.org/backup;\n -启动迁移后要耐心：根据您的Internet连接和硬件，升级所有内容最多可能需要几个小时。",
-    "migration_0015_system_not_fully_up_to_date": "您的系统不是最新的。请先执行常规升级，然后再运行向Buster的迁移。",
-    "migration_0015_not_enough_free_space": "/var/中的可用空间非常低！您应该至少有1GB的可用空间来运行此迁移。",
-    "migration_0015_not_stretch": "当前的Debian发行版不是Stretch！",
-    "migration_0015_yunohost_upgrade": "正在启动YunoHost核心升级...",
-    "migration_0015_still_on_stretch_after_main_upgrade": "在主要升级期间出了点问题，系统似乎仍在Debian Stretch上",
-    "migration_0015_main_upgrade": "正在开始主要升级...",
-    "migration_0015_patching_sources_list": "修补sources.lists ...",
-    "migration_0015_start": "开始迁移至Buster",
-    "migration_update_LDAP_schema": "正在更新LDAP模式...",
     "migration_ldap_rollback_success": "系统回滚。",
     "migration_ldap_migration_failed_trying_to_rollback": "无法迁移...试图回滚系统。",
     "migration_ldap_can_not_backup_before_migration": "迁移失败之前，无法完成系统的备份。错误: {error}",
     "migration_ldap_backup_before_migration": "在实际迁移之前，请创建LDAP数据库和应用程序设置的备份。",
-    "migration_description_0020_ssh_sftp_permissions": "添加SSH和SFTP权限支持",
-    "migration_description_0019_extend_permissions_features": "扩展/修改应用程序的权限管理系统",
-    "migration_description_0018_xtable_to_nftable": "将旧的网络流量规则迁移到新的nftable系统",
-    "migration_description_0017_postgresql_9p6_to_11": "将数据库从PostgreSQL 9.6迁移到11",
-    "migration_description_0016_php70_to_php73_pools": "将php7.0-fpm'pool'conf文件迁移到php7.3",
-    "migration_description_0015_migrate_to_buster": "将系统升级到Debian Buster和YunoHost 4.x",
-    "migrating_legacy_permission_settings": "正在迁移旧版权限设置...",
     "main_domain_changed": "主域已更改",
     "main_domain_change_failed": "无法更改主域",
     "mail_unavailable": "该电子邮件地址是保留的，并且将自动分配给第一个用户",
@@ -627,7 +591,7 @@
     "log_user_delete": "删除用户'{}'",
     "log_user_create": "添加用户'{}'",
     "domain_registrar_is_not_configured": "尚未为域 {domain} 配置注册商。",
-    "domain_dns_push_not_applicable": "的自动DNS配置的特征是不适用域{域}。您应该按照 https://yunohost.org/dns_config 上的文档手动配置DNS 记录。",
+    "domain_dns_push_not_applicable": "的自动DNS配置的特征是不适用域{domain}。您应该按照 https://yunohost.org/dns_config 上的文档手动配置DNS 记录。",
     "disk_space_not_sufficient_update": "没有足够的磁盘空间来更新此应用程序",
     "diagnosis_high_number_auth_failures": "最近出现了大量可疑的失败身份验证。您的fail2ban正在运行且配置正确，或使用自定义端口的SSH作为https://yunohost.org/解释的安全性。",
     "diagnosis_apps_not_in_app_catalog": "此应用程序不在 YunoHost 的应用程序目录中。如果它过去有被删除过，您应该考虑卸载此应用程，因为它不会更新，并且可能会损害您系统的完整和安全性。",
@@ -638,7 +602,7 @@
     "config_unknown_filter_key": "该过滤器钥匙“{filter_key}”有误。",
     "diagnosis_apps_outdated_ynh_requirement": "此应用程序的安装 版本只需要 yunohost >= 2.x，这往往表明它与推荐的打包实践和帮助程序不是最新的。你真的应该考虑更新它。",
     "disk_space_not_sufficient_install": "没有足够的磁盘空间来安装此应用程序",
-    "config_apply_failed": "应用新配置 失败：{错误}",
+    "config_apply_failed": "应用新配置 失败：{error}",
     "config_cant_set_value_on_section": "无法在整个配置部分设置单个值 。",
     "config_validate_color": "是有效的 RGB 十六进制颜色",
     "config_validate_date": "有效日期格式为YYYY-MM-DD",
