+++ maintenance/missing_i18n_keys.py
@@ -1,11 +1,17 @@
 # -*- coding: utf-8 -*-
 
+import toml
 import os
 import re
 import glob
 import json
 import yaml
 import subprocess
+import sys
+
+ROOT = os.path.dirname(__file__) + "/../"
+LOCALE_FOLDER = ROOT + "/locales/"
+REFERENCE_FILE = LOCALE_FOLDER + "en.json"
 
 ###############################################################################
 #   Find used keys in python code                                             #
@@ -24,12 +30,12 @@ def find_expected_string_keys():
     p3 = re.compile(r"YunohostValidationError\(\n*\s*[\'\"](\w+)[\'\"]")
     p4 = re.compile(r"# i18n: [\'\"]?(\w+)[\'\"]?")
 
-    python_files = glob.glob("src/*.py")
-    python_files.extend(glob.glob("src/utils/*.py"))
-    python_files.extend(glob.glob("src/migrations/*.py"))
-    python_files.extend(glob.glob("src/authenticators/*.py"))
-    python_files.extend(glob.glob("src/diagnosers/*.py"))
-    python_files.append("bin/yunohost")
+    python_files = glob.glob(ROOT + "src/*.py")
+    python_files.extend(glob.glob(ROOT + "src/utils/*.py"))
+    python_files.extend(glob.glob(ROOT + "src/migrations/*.py"))
+    python_files.extend(glob.glob(ROOT + "src/authenticators/*.py"))
+    python_files.extend(glob.glob(ROOT + "src/diagnosers/*.py"))
+    python_files.append(ROOT + "bin/yunohost")
 
     for python_file in python_files:
         content = open(python_file).read()
@@ -51,7 +57,9 @@ def find_expected_string_keys():
     # For each diagnosis, try to find strings like "diagnosis_stuff_foo" (c.f. diagnosis summaries)
     # Also we expect to have "diagnosis_description_<name>" for each diagnosis
     p3 = re.compile(r"[\"\'](diagnosis_[a-z]+_\w+)[\"\']")
-    for python_file in glob.glob("src/diagnosers/*.py"):
+    for python_file in glob.glob(ROOT + "src/diagnosers/*.py"):
+        if "__init__.py" in python_file:
+            continue
         content = open(python_file).read()
         for m in p3.findall(content):
             if m.endswith("_"):
@@ -63,14 +71,14 @@ def find_expected_string_keys():
         ]
 
     # For each migration, expect to find "migration_description_<name>"
-    for path in glob.glob("src/migrations/*.py"):
+    for path in glob.glob(ROOT + "src/migrations/*.py"):
         if "__init__" in path:
             continue
         yield "migration_description_" + os.path.basename(path)[:-3]
 
     # For each default service, expect to find "service_description_<name>"
     for service, info in yaml.safe_load(
-        open("conf/yunohost/services.yml")
+        open(ROOT + "conf/yunohost/services.yml")
     ).items():
         if info is None:
             continue
@@ -79,7 +87,7 @@ def find_expected_string_keys():
     # For all unit operations, expect to find "log_<name>"
     # A unit operation is created either using the @is_unit_operation decorator
     # or using OperationLogger(
-    cmd = "grep -hr '@is_unit_operation' src/ -A3 2>/dev/null | grep '^def' | sed -E 's@^def (\\w+)\\(.*@\\1@g'"
+    cmd = f"grep -hr '@is_unit_operation' {ROOT}/src/ -A3 2>/dev/null | grep '^def' | sed -E 's@^def (\\w+)\\(.*@\\1@g'"
     for funcname in (
         subprocess.check_output(cmd, shell=True).decode("utf-8").strip().split("\n")
     ):
@@ -94,14 +102,14 @@ def find_expected_string_keys():
     # Global settings descriptions
     # Will be on a line like : ("service.ssh.allow_deprecated_dsa_hostkey", {"type": "bool", ...
     p5 = re.compile(r" \(\n*\s*[\"\'](\w[\w\.]+)[\"\'],")
-    content = open("src/settings.py").read()
+    content = open(ROOT + "src/settings.py").read()
     for m in (
         "global_settings_setting_" + s.replace(".", "_") for s in p5.findall(content)
     ):
         yield m
 
     # Keys for the actionmap ...
-    for category in yaml.safe_load(open("share/actionsmap.yml")).values():
+    for category in yaml.safe_load(open(ROOT + "share/actionsmap.yml")).values():
         if "actions" not in category.keys():
             continue
         for action in category["actions"].values():
@@ -129,65 +137,72 @@ def find_expected_string_keys():
         yield "backup_applying_method_%s" % method
         yield "backup_method_%s_finished" % method
 
-    for level in ["danger", "thirdparty", "warning"]:
-        yield "confirm_app_install_%s" % level
+    registrars = toml.load(open(ROOT + "share/registrar_list.toml"))
+    supported_registrars = ["ovh", "gandi", "godaddy"]
+    for registrar in supported_registrars:
+        for key in registrars[registrar].keys():
+            yield f"domain_config_{key}"
 
-    for errortype in ["not_found", "error", "warning", "success", "not_found_details"]:
-        yield "diagnosis_domain_expiration_%s" % errortype
-    yield "diagnosis_domain_not_found_details"
-
-    for errortype in ["bad_status_code", "connection_error", "timeout"]:
-        yield "diagnosis_http_%s" % errortype
-
-    yield "password_listed"
-    for i in [1, 2, 3, 4]:
-        yield "password_too_simple_%s" % i
-
-    checks = [
-        "outgoing_port_25_ok",
-        "ehlo_ok",
-        "fcrdns_ok",
-        "blacklist_ok",
-        "queue_ok",
-        "ehlo_bad_answer",
-        "ehlo_unreachable",
-        "ehlo_bad_answer_details",
-        "ehlo_unreachable_details",
-    ]
-    for check in checks:
-        yield "diagnosis_mail_%s" % check
-
-
-###############################################################################
-#   Load en locale json keys                                                  #
-###############################################################################
-
-
-def keys_defined_for_en():
-    return json.loads(open("locales/en.json").read()).keys()
+    domain_config = toml.load(open(ROOT + "share/config_domain.toml"))
+    for panel in domain_config.values():
+        if not isinstance(panel, dict):
+            continue
+        for section in panel.values():
+            if not isinstance(section, dict):
+                continue
+            for key, values in section.items():
+                if not isinstance(values, dict):
+                    continue
+                yield f"domain_config_{key}"
 
 
 ###############################################################################
 #   Compare keys used and keys defined                                        #
 ###############################################################################
 
+if len(sys.argv) <= 1 or sys.argv[1] not in ["--check", "--fix"]:
+    print("Please specify --check or --fix")
+    sys.exit(1)
 
 expected_string_keys = set(find_expected_string_keys())
-keys_defined = set(keys_defined_for_en())
+keys_defined_for_en = json.loads(open(REFERENCE_FILE).read()).keys()
+keys_defined = set(keys_defined_for_en)
 
+unused_keys = keys_defined.difference(expected_string_keys)
+unused_keys = sorted(unused_keys)
 
 undefined_keys = expected_string_keys.difference(keys_defined)
 undefined_keys = sorted(undefined_keys)
 
+mode = sys.argv[1].strip("-")
+if mode == "check":
 
-j = json.loads(open("locales/en.json").read())
-for key in undefined_keys:
-    j[key] = "FIXME"
+    # Unused keys are not too problematic, will be automatically
+    # removed by the other autoreformat script,
+    # but still informative to display them
+    if unused_keys:
+        print(
+            "Those i18n keys appears unused:\n" "    - " + "\n    - ".join(unused_keys)
+        )
 
-json.dump(
-    j,
-    open("locales/en.json", "w"),
-    indent=4,
-    ensure_ascii=False,
-    sort_keys=True,
-)
+    if undefined_keys:
+        print(
+            "Those i18n keys should be defined in en.json:\n"
+            "    - " + "\n    - ".join(undefined_keys)
+        )
+        sys.exit(1)
+
+elif mode == "fix":
+    j = json.loads(open(REFERENCE_FILE).read())
+    for key in undefined_keys:
+        j[key] = "FIXME"
+    for key in unused_keys:
+        del j[key]
+
+    json.dump(
+        j,
+        open(REFERENCE_FILE, "w"),
+        indent=4,
+        ensure_ascii=False,
+        sort_keys=True,
+    )
